Rename implementation.cpp to implementation2.cpp
I've updated the existing implementation.cpp to display the score when you win (not just to display when you die).  Also in this includes some effect that stamina will have on your character.  In this case the amount of health lost when you run out of food is reduced the more stamina you have.

I was also thinking you could integrate luck into the amount paid for food and gas.